+++ core-web/libs/block-editor/src/lib/extensions/asset-form/asset-form.component.html
@@ -28,6 +28,7 @@
                         [type]="type"
                         (uploadedFile)="onSelectAsset($event)"
                         (preventClose)="onPreventClose($event)"
+                        (hide)="onHide($event)"
                     ></dot-upload-asset>
                 </div>
             </ng-template>
